integration: "testV3WatchMultipleStreams" with stream coalescing
Now that watch response returns the original watch IDs,
gRPC proxy test should expect coalesced watch ID,
while regular test expects unique watch IDs.
