+++ integration/v3_watch_test.go
@@ -864,112 +864,6 @@ func TestV3WatchMultipleEventsPutUnsynced(t *testing.T) {
 	}
 }
 
-func TestV3WatchMultipleStreamsSynced(t *testing.T) {
-	defer testutil.AfterTest(t)
-	testV3WatchMultipleStreams(t, 0)
-}
-
-func TestV3WatchMultipleStreamsUnsynced(t *testing.T) {
-	defer testutil.AfterTest(t)
-	testV3WatchMultipleStreams(t, 1)
-}
-
-// testV3WatchMultipleStreams tests multiple watchers on the same key on multiple streams.
-func testV3WatchMultipleStreams(t *testing.T, startRev int64) {
-	clus := NewClusterV3(t, &ClusterConfig{Size: 3})
-	defer clus.Terminate(t)
-
-	wAPI := toGRPC(clus.RandClient()).Watch
-	kvc := toGRPC(clus.RandClient()).KV
-
-	streams := make([]pb.Watch_WatchClient, 5)
-	for i := range streams {
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		wStream, errW := wAPI.Watch(ctx)
-		if errW != nil {
-			t.Fatalf("wAPI.Watch error: %v", errW)
-		}
-		wreq := &pb.WatchRequest{RequestUnion: &pb.WatchRequest_CreateRequest{
-			CreateRequest: &pb.WatchCreateRequest{
-				Key: []byte("foo"), StartRevision: startRev}}}
-		if err := wStream.Send(wreq); err != nil {
-			t.Fatalf("wStream.Send error: %v", err)
-		}
-		streams[i] = wStream
-	}
-
-	for _, wStream := range streams {
-		wresp, err := wStream.Recv()
-		if err != nil {
-			t.Fatalf("wStream.Recv error: %v", err)
-		}
-		if !wresp.Created {
-			t.Fatalf("wresp.Created got = %v, want = true", wresp.Created)
-		}
-	}
-
-	if _, err := kvc.Put(context.TODO(), &pb.PutRequest{Key: []byte("foo"), Value: []byte("bar")}); err != nil {
-		t.Fatalf("couldn't put key (%v)", err)
-	}
-
-	var wg sync.WaitGroup
-	wg.Add(len(streams))
-	wevents := []*mvccpb.Event{
-		{
-			Type: mvccpb.PUT,
-			Kv:   &mvccpb.KeyValue{Key: []byte("foo"), Value: []byte("bar"), CreateRevision: 2, ModRevision: 2, Version: 1},
-		},
-	}
-	for i := range streams {
-		go func(i int) {
-			defer wg.Done()
-			wStream := streams[i]
-			wresp, err := wStream.Recv()
-			if err != nil {
-				t.Fatalf("wStream.Recv error: %v", err)
-			}
-			if wresp.WatchId != 0 {
-				t.Errorf("watchId got = %d, want = 0", wresp.WatchId)
-			}
-			if !reflect.DeepEqual(wresp.Events, wevents) {
-				t.Errorf("wresp.Events got = %+v, want = %+v", wresp.Events, wevents)
-			}
-			// now Recv should block because there is no more events coming
-			rok, nr := waitResponse(wStream, 1*time.Second)
-			if !rok {
-				t.Errorf("unexpected pb.WatchResponse is received %+v", nr)
-			}
-		}(i)
-	}
-	wg.Wait()
-}
-
-// waitResponse waits on the given stream for given duration.
-// If there is no more events, true and a nil response will be
-// returned closing the WatchClient stream. Or the response will
-// be returned.
-func waitResponse(wc pb.Watch_WatchClient, timeout time.Duration) (bool, *pb.WatchResponse) {
-	rCh := make(chan *pb.WatchResponse, 1)
-	donec := make(chan struct{})
-	defer close(donec)
-	go func() {
-		resp, _ := wc.Recv()
-		select {
-		case rCh <- resp:
-		case <-donec:
-		}
-	}()
-	select {
-	case nr := <-rCh:
-		return false, nr
-	case <-time.After(timeout):
-	}
-	// didn't get response
-	wc.CloseSend()
-	return true, nil
-}
-
 func TestWatchWithProgressNotify(t *testing.T) {
 	// accelerate report interval so test terminates quickly
 	oldpi := v3rpc.GetProgressReportInterval()
